requirements.txt
# Core
streamlit>=1.30.0
plotly>=5.18.0
kaleido>=0.2.1
biopython>=1.83
pandas>=2.1.0
numpy>=1.26.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# PDF Report Generation
fpdf2>=2.7.0

# Authentication
streamlit-authenticator>=0.3.1
PyYAML>=6.0

# REST API
flask>=3.0.0
flask-limiter>=3.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting
flake8>=7.0.0
